further split of load_data
